ISPN-14898 Moved string tests on new file
